+++ CHANGELOG.md
@@ -10,6 +10,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - A `disconnect` signal to let Infinity know that the `MediaConnection` is going away
+- `ConferenceStep.conference(String)` to create a new `ConferenceStep` using a different alias
 
 ### Changed